+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Features:
 
 - xpiler: now supports wildcard search pattern matching in arguments (not regex)
 - introduced builtin event LocalEvent
+- ServerLink: added methods GetSession(s)
 
 Bugfixes: